Split graphql types into different conxtexts
