+++ lib/zssn_web/resolvers/survivors.ex
@@ -1,0 +1,8 @@
+defmodule ZssnWeb.Resolvers.Survivors do
+  alias Zssn.Graphql.Survivors
+  alias Zssn.Survivors.Survivor
+
+  def survivors(_, args, _) do
+    {:ok, Survivors.list(Survivor, args)}
+  end
+end
